+++ scm-common-data/src/main/java/ir/daneshrefah/scm/common/data/repository/assets/MembershipTerminalServiceAccessRepository.java
@@ -31,10 +31,10 @@ public interface MembershipTerminalServiceAccessRepository extends JpaRepository
             "channelServiceAccess",
             "channelServiceAccess.ebService"
     })
-    List<MembershipTerminalServiceAccessEntity> findByLegacyTerminalIdAndAccountNo(@Param("channelId") Short channelId, @Param("accountNo") String accountNo,@Param("personId") Long personId);
+    List<MembershipTerminalServiceAccessEntity> findByLegacyTerminalIdAndAccountNo(@Param("channelId") Short channelId, @Param("accountNo") String accountNo,@Param("personId") Integer personId);
 
 
     @Query("select o from MembershipTerminalServiceAccessEntity o where o.channelServiceAccess.id = :channelServiceAccessId and o.membershipTerminalAccess.membership.customerAccount.account.accountNo = :accountNo and o.membershipTerminalAccess.membership.person.id = :personId")
-    List<MembershipTerminalServiceAccessEntity> findByChannelServiceAccessIdAndAccountNoAndPersonId(@Param("channelServiceAccessId") Long channelServiceAccessId, @Param("accountNo") String accountNo,@Param("personId") Long personId);
+    List<MembershipTerminalServiceAccessEntity> findByChannelServiceAccessIdAndAccountNoAndPersonId(@Param("channelServiceAccessId") Long channelServiceAccessId, @Param("accountNo") String accountNo,@Param("personId") Integer personId);
 
 }
