+++ scm-common-data/src/main/java/ir/daneshrefah/scm/common/data/service/person/PersonService.java
@@ -1,9 +1,11 @@
 package ir.daneshrefah.scm.common.data.service.person;
 
-import ir.daneshrefah.scm.common.data.entity.person.ClientPersonEntity;
 import ir.daneshrefah.scm.common.dto.membership.PersonFindRequest;
 import ir.daneshrefah.scm.common.dto.spec.PagedResponseData;
-import ir.daneshrefah.scm.common.model.person.*;
+import ir.daneshrefah.scm.common.model.person.ClientPerson;
+import ir.daneshrefah.scm.common.model.person.GeneralPerson;
+import ir.daneshrefah.scm.common.model.person.GeneralRealPerson;
+import ir.daneshrefah.scm.common.model.person.PersonType;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,7 +23,7 @@ public interface PersonService {
 
     boolean checkPersonExist(PersonFindRequest request);
 
-    GeneralPerson findPersonByPersonId(Long id);
+    GeneralPerson findPersonByPersonId(Integer id);
 
     Optional<GeneralPerson> findPersonByUsername(String username);
